move settings panel to ui
